fix: move contracts page tests to correct location and remove unused import
