+++ src/components/__tests__/ContractCreationForm.test.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import { render, screen, fireEvent, waitFor } from '@testing-library/react';
-import userEvent from '@testing-library/user-event';
 import '@testing-library/jest-dom';
 import { ContractCreationForm } from '../ContractCreationForm';
 import * as stellarAddress from '@/lib/stellarAddress';
